refactor: move session related objects to session dir
