+++ src/Session/Objects/Objects.php
@@ -1,0 +1,12 @@
+<?php
+namespace Give\Session\Objects;
+
+interface Objects {
+	/**
+	 * Take array return object
+	 *
+	 * @since 2.7.0
+	 * @param array $array
+	 */
+	public static function fromArray( $array );
+}
